fix(reddog): close artifact model authority lineage
Worker-Lane: 0102-architect

Slice: REDDOG_ARTIFACT_GENERATION_BOUND_MODEL_RUNTIME_PHASE1
