+++ extensions/reddog/backend_compatibility_constants.js
@@ -4,7 +4,7 @@ const BACKEND_MANIFEST_SCHEMA = 'reddog_backend_manifest.v3';
 const BACKEND_PRODUCT = 'foundups-agent-reddog-backend';
 const BACKEND_API_VERSION = 2;
 const BACKEND_MANIFEST_PATH = 'scripts/reddog_backend_manifest.json';
-const EXPECTED_MANIFEST_SHA256 = 'ab9d0878ebfa7902ed329409dd3e2dd43ba463e3602b205c2557ab82ecd473d3';
+const EXPECTED_MANIFEST_SHA256 = 'e1adbecfc84005ae9a0c4b310189bafefc6d0d8a2c49a0012a52c71a5e3645f6';
 const RUNTIME_DEPENDENCY_GRAPH_VERSION = 2;
 const MAX_MANIFEST_BYTES = 256 * 1024;
 const MAX_RUNTIME_FILE_BYTES = 2 * 1024 * 1024;
