+++ modules/ai_intelligence/ai_gateway/src/model_panel_signed_evidence.py
@@ -30,6 +30,7 @@ from .model_intelligence_catalog import ModelCatalogSnapshot
 from .model_intelligence_outcomes import ModelBenchmarkEvidenceReceipt, ModelPromotionEvidenceReceipt
 from .model_intelligence_selection import ModelSelectionReceipt, SelectionDecision, SelectionMode, SelectionPurpose
 from .model_runtime_binding import ModelRuntimeBindingPolicy
+from . import model_evidence_authority_validation as evidence_authority
 from .model_signed_evidence import (
     ModelEvidenceKeyResolver,
     ModelSignedEvidenceReceipt,
@@ -37,7 +38,6 @@ from .model_signed_evidence import (
     VerifiedModelProductionEvidence,
     build_verified_model_production_evidence,
 )
-
 
 PANEL_EVIDENCE_SCHEMA_VERSION = "model_panel_signed_evidence_receipt.v1"
 VERIFIED_PANEL_EVIDENCE_SCHEMA_VERSION = "verified_model_panel_evidence.v1"
@@ -412,7 +412,7 @@ def _verify_model_panel_evidence_inputs(
         leeway_s=leeway_s,
     )
     receipt = rehydrate_model_panel_signed_evidence_receipt(aggregate_record)
-    _assert_panel_signer_independence(receipt, entries)
+    evidence_authority.assert_independent_panel_authority(receipt, entries)
     _assert_panel_context(
         catalog_snapshot=catalog_snapshot,
         selection_receipt=selection_receipt,
@@ -435,33 +435,6 @@ def _verify_model_panel_evidence_inputs(
     )
     _consume_panel_nonce(receipt, nonce_store=nonce_store, consume_nonce=consume_nonce)
     return receipt, entries
-
-
-def _assert_panel_signer_independence(
-    receipt: ModelPanelSignedEvidenceReceipt,
-    entries: tuple[VerifiedModelEvidenceEntry, ...],
-) -> None:
-    member_keys = {
-        signature.signer_public_key
-        for entry in entries
-        for signature in (
-            entry.benchmark_signature_receipt,
-            entry.promotion_signature_receipt,
-        )
-    }
-    member_fingerprints = {
-        signature.signer_key_fingerprint
-        for entry in entries
-        for signature in (
-            entry.benchmark_signature_receipt,
-            entry.promotion_signature_receipt,
-        )
-    }
-    if (
-        receipt.signer_public_key in member_keys
-        or receipt.signer_key_fingerprint in member_fingerprints
-    ):
-        raise ValueError("panel_authority_signer_not_independent")
 
 
 build_verified_model_panel_evidence, _verified_panel_seal = _build_verified_panel_registry(
