+++ modules/communication/moltbot_bridge/tests/test_reddog_main_architect_fix_promotion_bootstrap.py
@@ -40,13 +40,6 @@ from modules.communication.moltbot_bridge.tests.architect_proposal_test_helpers 
     ready_proposal_policy,
 )
 from modules.infrastructure.database.src.db_manager import DatabaseManager
-from modules.ai_intelligence.ai_gateway.src.model_runtime_binding_verified_admission import (
-    verified_runtime_binding_receipt,
-)
-from modules.communication.moltbot_bridge.tests.model_runtime_binding_receipt_test_helpers import (
-    model_runtime_binding_test_capability,
-    model_selection_and_runtime_binding_receipts,
-)
 
 
 REPO_ROOT = Path(__file__).resolve().parents[4]
@@ -509,61 +502,6 @@ def test_crash_after_publication_reclaims_and_reconstructs_exact_promotion(
     assert row["status"] == "APPLIED"
     assert row["promotion_receipt_id"] == second.promotion_receipt_id
     assert row["committed_revision"] == second.committed_revision
-
-
-def test_bootstrap_forwards_runtime_binding_receipt_into_promotion(tmp_path: Path) -> None:
-    repo = _repo(tmp_path)
-    files = _runtime_files(tmp_path)
-    selection, binding = model_selection_and_runtime_binding_receipts(
-        runtime_surface="reddog_artifact_generation",
-        task_family="reddog_architect_fix_promotion",
-    )
-    files["model_selection"].write_text(
-        json.dumps(selection, sort_keys=True), encoding="utf-8"
-    )
-    files["model_runtime_binding"].write_text(
-        json.dumps(binding, sort_keys=True), encoding="utf-8"
-    )
-    verification = verified_runtime_binding_receipt(binding)
-    assert verification is not None
-    capability = model_runtime_binding_test_capability(selection, binding)
-
-    with (
-        patch(
-            "modules.communication.moltbot_bridge.src."
-            "reddog_main_architect_fix_promotion_bootstrap.read_git_head_sha",
-            return_value="sha256:repo-head",
-        ),
-        patch(
-            "modules.communication.moltbot_bridge.src."
-            "reddog_architect_proposal_admission_contract."
-            "current_architect_proposal_admission_policy",
-            return_value=ready_proposal_policy(),
-        ),
-    ):
-        result = run_reddog_main_architect_fix_promotion_bootstrap(
-            repo_root=repo,
-            runtime_root=tmp_path / "runtime",
-            work_state_path=files["work_state"],
-            architect_determination_path=files["determination"],
-            model_selection_receipt_path=files["model_selection"],
-            model_runtime_binding_receipt_path=files["model_runtime_binding"],
-            model_runtime_binding_verification_capability=capability,
-            memex_supply_receipt_path=files["memex_supply"],
-            authority_profile_source_path=files["authority_profile_source"],
-            authority_profile_output_path=files["authority_profile_output"],
-            holoindex_receipt_path=files["holoindex_receipt"],
-            worker_id="reddog-main-test",
-            now_iso=NOW,
-        )
-
-    assert result.accepted is True, result.rejection_reasons
-    promoted_profile = json.loads(files["authority_profile_output"].read_text(encoding="utf-8"))
-    runtime_binding = json.loads(files["model_runtime_binding"].read_text(encoding="utf-8"))
-    assert promoted_profile["model_runtime_binding_receipt_id"] == runtime_binding["receipt_id"]
-    assert promoted_profile["operational_context_binding"]["model_runtime_binding_receipt_id"] == (
-        runtime_binding["receipt_id"]
-    )
 
 
 def test_bootstrap_rejects_authority_profile_output_inside_repo(tmp_path: Path) -> None:
