+++ modules/communication/moltbot_bridge/tests/test_reddog_openclaw_hermes_0102_worker_dispatch_runtime.py
@@ -37,6 +37,9 @@ from modules.communication.moltbot_bridge.tests.reddog_resident_queue_test_helpe
     worker_dispatch_authority_stages,
     with_architect_fix_publication,
 )
+from modules.communication.moltbot_bridge.tests.model_runtime_binding_queue_test_helpers import (
+    runtime_binding_refs as _runtime_binding_refs,
+)
 
 
 REPO_ROOT = Path(__file__).resolve().parents[4]
@@ -278,17 +281,6 @@ def _publish(**kwargs):
         worker_dispatch_authority_verification_context(),
     )
     return runtime.publish_reddog_signed_worker_dispatch_runtime(**kwargs)
-
-
-def _runtime_binding_refs():
-    return {
-        "model_runtime_binding_receipt_id": "reddog_model_runtime_binding:abc123",
-        "model_runtime_binding_digest": "sha256:" + "a" * 64,
-        "model_runtime_binding_verification_receipt_id": (
-            "model_runtime_binding_verification:" + "b" * 64
-        ),
-        "model_runtime_binding_verification_digest": "sha256:" + "c" * 64,
-    }
 
 
 def _memex_refs():
@@ -1159,12 +1151,7 @@ def test_rejects_model_runtime_binding_conflict_between_signed_receipt_and_queue
         worker_dispatch_dryrun_result=dryrun,
         work_state_snapshot=_snapshot(
             allocation,
-            **{
-                **refs,
-                "model_runtime_binding_receipt_id": (
-                    "reddog_model_runtime_binding:other"
-                ),
-            },
+            **{**refs, "model_runtime_binding_receipt_id": "reddog_model_runtime_binding:other"},
         ),
         queue_item_id="queue-1",
         writer=_FakeWriter(),
