+++ modules/communication/moltbot_bridge/tests/test_reddog_resident_queue_model_runtime_artifact_integration.py
@@ -20,6 +20,12 @@ from modules.communication.moltbot_bridge.src import (
 from modules.communication.moltbot_bridge.src.reddog_authority_profile_safety import (
     authority_profile_unknown_field_paths,
 )
+from modules.communication.moltbot_bridge.src.reddog_architect_fix_promotion_publication import (
+    AtomicArchitectFixPromotionPublisher,
+)
+from modules.communication.moltbot_bridge.src.reddog_authoritative_work_state_refresh_runtime import (
+    InMemoryAuthoritativeWorkStateStore,
+)
 from modules.communication.moltbot_bridge.src.reddog_bounded_artifact_generation_runtime import (
     FoundupsFusionArtifactGenerationRunner,
 )
@@ -32,6 +38,9 @@ from modules.communication.moltbot_bridge.src.reddog_resident_queue_bounded_work
 )
 from modules.communication.moltbot_bridge.src.reddog_resident_queue_next_stage_dispatch import (
     invoke_reddog_resident_queue_next_stage_dispatch,
+)
+from modules.communication.moltbot_bridge.src.reddog_resident_queue_chain_results_store import (
+    AtomicJsonResidentQueueChainResultsStore,
 )
 from modules.communication.moltbot_bridge.tests.model_runtime_binding_queue_test_helpers import (
     model_bound_profile,
@@ -59,7 +68,6 @@ from modules.communication.moltbot_bridge.tests.test_reddog_main_resident_queue_
     _pilot_path_overrides,
     _pilot_bounded_worker_plan,
     _pilot_allowed_paths,
-    _pilot_signed_authority_for_bootstrap,
     _pilot_payloads,
     _pilot_worktree_path,
     _principals,
@@ -72,7 +80,6 @@ from modules.communication.moltbot_bridge.tests.test_reddog_main_resident_queue_
 from modules.communication.moltbot_bridge.tests.test_reddog_resident_queue_bounded_worker_pilot_handler import (
     _Resolver,
     _binding_stage_overrides,
-    _mapping_digest as _handler_mapping_digest,
     _seeded_store,
     _snapshot as _handler_snapshot,
     _valid_bundle,
@@ -83,6 +90,11 @@ from modules.communication.moltbot_bridge.tests.test_reddog_architect_fix_signed
     _determination,
     _promote,
     _rebind_determination_admission,
+    _work_state,
+)
+from modules.communication.moltbot_bridge.tests.architect_proposal_promotion_test_helpers import (
+    PRINCIPAL_PRIVATE_KEY,
+    REDDOG_PRIVATE_KEY,
 )
 from modules.communication.moltbot_bridge.tests.test_reddog_wre_queue_authorized_bounded_worker_pilot_invoke import (
     ARTIFACT as HANDLER_ARTIFACT,
@@ -93,7 +105,10 @@ def test_bootstrap_generates_model_bound_artifacts_before_pilot(
     tmp_path: Path,
 ) -> None:
     repo = _repo(tmp_path)
-    principal_public, reddog_public, connector = _ed25519_signing_material()
+    principal_public, reddog_public, connector = _ed25519_signing_material(
+        principal_key=PRINCIPAL_PRIVATE_KEY,
+        reddog_key=REDDOG_PRIVATE_KEY,
+    )
     selection, binding = model_selection_and_runtime_binding_receipts(
         runtime_surface=RUNTIME_SURFACE_ARTIFACT_GENERATION
     )
@@ -214,15 +229,64 @@ def test_promoted_queue_claim_materialization_reaches_exact_provider(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    binding, snapshot, queue_item, worker_claim, work_order = (
-        _promote_claimed_work_order(monkeypatch)
+    context = _promoted_runtime_context(tmp_path, monkeypatch)
+    provider_calls = _invoke_promoted_provider(context, monkeypatch)
+    assert len(provider_calls) == 1
+    _assert_continuous_lineage(
+        context["queue_item"],
+        context["worker_claim"],
+        context["work_order"],
+        provider_calls[0],
+        context["chain_store"].load(),
     )
-    bundle = _valid_bundle(tmp_path)
-    chain_store = _promoted_chain_store(
-        bundle,
-        queue_item=queue_item,
-        work_order=work_order,
+
+
+def _promoted_runtime_context(tmp_path, monkeypatch):
+    repo = _repo(tmp_path)
+    principal_public, reddog_public, connector = _ed25519_signing_material(
+        principal_key=PRINCIPAL_PRIVATE_KEY,
+        reddog_key=REDDOG_PRIVATE_KEY,
     )
+    binding, snapshot, queue_item, worker_claim, work_order, authority_profile = (
+        _promote_claimed_work_order(
+            monkeypatch,
+            tmp_path=tmp_path,
+            repo=repo,
+        )
+    )
+    state = _write_runtime_json(tmp_path, "promoted_work_state.json", snapshot)
+    profile = _write_runtime_json(
+        tmp_path, "promoted_authority_profile.json", authority_profile
+    )
+    paths = _runtime_paths(tmp_path, repo, work_order, principal_public)
+    verifier = model_runtime_binding_test_verifier(binding)
+    assert verifier is not None
+    bootstrap = _run_bootstrap(
+        repo,
+        state,
+        profile,
+        paths,
+        connector,
+        _FakeArtifactGenerator(content="# bootstrap must not generate\n"),
+        requested_queue_item_id=queue_item["queue_item_id"],
+        model_runtime_binding_verifier=verifier,
+    )
+    _assert_bootstrap_yielded_at_assurance(bootstrap, paths["chain"])
+    chain_store = AtomicJsonResidentQueueChainResultsStore(
+        paths["chain"], allowed_root=tmp_path
+    )
+    return {
+        "binding": binding,
+        "snapshot": snapshot,
+        "queue_item": queue_item,
+        "worker_claim": worker_claim,
+        "work_order": work_order,
+        "repo": repo,
+        "chain_store": chain_store,
+    }
+
+
+def _invoke_promoted_provider(context, monkeypatch):
     provider_calls: list[dict[str, object]] = []
     monkeypatch.setenv("OPENROUTER_API_KEY", "test-key")
     monkeypatch.setattr(
@@ -231,35 +295,35 @@ def test_promoted_queue_claim_materialization_reaches_exact_provider(
         lambda: _provider_stub(provider_calls, PILOT_ARTIFACT),
     )
     handler = build_reddog_resident_queue_bounded_worker_pilot_stage_handler(
-        chain_results_store=chain_store,
-        work_order_resolver=_Resolver(work_order),
+        chain_results_store=context["chain_store"],
+        work_order_resolver=_Resolver(context["work_order"]),
         artifact_contents={},
         artifact_generation_request_binding_enabled=True,
         artifact_generator=FoundupsFusionArtifactGenerationRunner(
             runtime_mode="foundups_fusion"
         ),
-        model_runtime_binding_verifier=model_runtime_binding_test_verifier(binding),
-        repo_root=bundle["repo_root"],
+        model_runtime_binding_verifier=model_runtime_binding_test_verifier(
+            context["binding"]
+        ),
+        repo_root=context["repo"],
     )
     result = invoke_reddog_resident_queue_next_stage_dispatch(
         explicit_resident_queue_stage_dispatch_requested=True,
-        work_state_snapshot=snapshot,
-        store=chain_store,
+        work_state_snapshot=context["snapshot"],
+        store=context["chain_store"],
         handlers={BOUNDED_WORKER_PILOT_STAGE_KEY: handler},
         now_iso=NOW,
     )
     assert result.accepted is True, result.rejection_reasons
-    assert len(provider_calls) == 1
-    _assert_continuous_lineage(
-        queue_item,
-        worker_claim,
-        work_order,
-        provider_calls[0],
-        chain_store.load(),
-    )
+    return provider_calls
 
 
-def _promote_claimed_work_order(monkeypatch):
+def _promote_claimed_work_order(
+    monkeypatch,
+    *,
+    tmp_path: Path,
+    repo: Path,
+):
     selection, binding = model_selection_and_runtime_binding_receipts(
         runtime_surface=RUNTIME_SURFACE_ARTIFACT_GENERATION,
         task_family="reddog_architect_fix_promotion",
@@ -277,11 +341,23 @@ def _promote_claimed_work_order(monkeypatch):
         "_work_order_id",
         lambda _queue_item_id: WORK_ORDER_ID,
     )
+    store = InMemoryAuthoritativeWorkStateStore(_work_state())
+    publication_runtime = tmp_path / "publication_runtime"
+    publisher = AtomicArchitectFixPromotionPublisher(
+        repo_root=repo,
+        runtime_root=publication_runtime,
+        authority_profile_path=publication_runtime / "authority_profile.json",
+        work_state_store=store,
+    )
     promoted, store = _promote(
+        store=store,
         model_selection_receipt=selection,
         model_runtime_binding_receipt=binding,
-        authority_profile=_promoted_authority_profile(_promoted_bounded_worker_plan()),
+        authority_profile=_promoted_authority_profile(
+            _promoted_bounded_worker_plan(),
+        ),
         architect_determination=determination,
+        authority_profile_publication_publisher=publisher.publish,
     )
     assert promoted.accepted is True, promoted.rejection_reasons
     snapshot = store.load()
@@ -294,7 +370,14 @@ def _promote_claimed_work_order(monkeypatch):
     )
     assert reasons == ()
     assert work_orders is not None
-    return binding, snapshot, queue_item, snapshot["worker_claims"][0], work_orders[WORK_ORDER_ID]
+    return (
+        binding,
+        snapshot,
+        queue_item,
+        snapshot["worker_claims"][0],
+        work_orders[WORK_ORDER_ID],
+        promoted.authority_profile,
+    )
 
 
 def _promoted_bounded_worker_plan():
@@ -303,13 +386,16 @@ def _promoted_bounded_worker_plan():
     return plan
 
 
-def _promoted_authority_profile(plan):
+def _promoted_authority_profile(
+    plan,
+):
     path_overrides = _pilot_path_overrides()
     path_overrides.pop("task_summary", None)
     path_overrides.pop("rollback_plan", None)
     return _authority_profile(
         foundup_id=PILOT_DOMAIN_ID,
         permission_snapshot_digest=PERMISSION_DIGEST,
+        sovereign_authorization_digest="sha256:" + ("e" * 64),
         holoindex_evidence={
             "holoindex_query": "Promoted pAccess bounded artifact generation",
             "holoindex_status": "bundle_json_ok",
@@ -327,38 +413,15 @@ def _promoted_authority_profile(plan):
     )
 
 
-def _promoted_chain_store(bundle, *, queue_item, work_order):
-    overrides = _binding_stage_overrides()
-    authority_result = overrides["authority_runtime"]["authority_result"]
-    authority = authority_result["work_authority"]
-    authority.clear()
-    authority.update(_pilot_signed_authority_for_bootstrap())
-    for field in _LINEAGE_FIELDS:
-        authority[field] = work_order[field]
-    authority_digest = _handler_mapping_digest(authority)
-    authority_result["receipt"]["work_authority_digest"] = authority_digest
-    overrides["authority_verification"][
-        "verified_work_authority_digest"
-    ] = authority_digest
-    seeded = _seeded_store(bundle, **overrides)
-    payload = seeded.load()
-    payload["queue_item_id"] = queue_item["queue_item_id"]
-    payload["selected_slice"] = queue_item["slice_id"]
-    worktree = payload["stage_results"]["worktree_create"][
-        "worktree_create_result"
-    ]
-    worktree["work_order_id"] = work_order["work_order_id"]
-    worktree["branch_name"] = work_order["branch_name"]
-    return seeded.__class__(payload)
-
-
 def _assert_continuous_lineage(queue_item, worker_claim, work_order, call, chain):
     expected = {field: work_order[field] for field in _LINEAGE_FIELDS}
     assert {field: queue_item[field] for field in _LINEAGE_FIELDS} == expected
-    for field in _CLAIM_LINEAGE_FIELDS:
-        assert worker_claim[field] == expected[field]
+    assert {field: worker_claim[field] for field in _LINEAGE_FIELDS} == expected
     bridge_meta = call["bridge_meta"]
     assert {field: bridge_meta[field] for field in _LINEAGE_FIELDS} == expected
+    runtime = work_order["model_runtime_binding_receipt"]
+    assert call["lead_model"] == runtime["principal_model"]
+    assert call["panel_models"] == runtime["panel_models"]
     generation = chain["stage_results"][BOUNDED_WORKER_PILOT_STAGE_KEY][
         "artifact_generation_result"
     ]["receipt"]
@@ -388,14 +451,19 @@ _LINEAGE_FIELDS = (
     "model_runtime_binding_verification_receipt_id",
     "model_runtime_binding_verification_digest",
 )
-_CLAIM_LINEAGE_FIELDS = (
-    "model_selection_receipt_id",
-    "model_runtime_binding_receipt_id",
-    "model_runtime_binding_verification_receipt_id",
-)
 
 
-def _run_bootstrap(repo, state, profile, paths, connector, artifact_generator):
+def _run_bootstrap(
+    repo,
+    state,
+    profile,
+    paths,
+    connector,
+    artifact_generator,
+    *,
+    requested_queue_item_id="queue-1",
+    model_runtime_binding_verifier=None,
+):
     return run_reddog_main_resident_queue_serial_loop_bootstrap(
         repo_root=repo,
         work_state_path=state,
@@ -416,11 +484,13 @@ def _run_bootstrap(repo, state, profile, paths, connector, artifact_generator):
         worker_dispatch_writer=_FakeWorkerDispatchTaskWriter(),
         worktree_runner=_FakeWorktreeRunner(),
         artifact_generator=artifact_generator,
-        model_runtime_binding_verifier=_FakeModelRuntimeBindingVerifier(),
+        model_runtime_binding_verifier=(
+            model_runtime_binding_verifier or _FakeModelRuntimeBindingVerifier()
+        ),
         now_iso=NOW,
         now_epoch=1000,
         trusted_now_epoch=lambda: 1000,
-        requested_queue_item_id="queue-1",
+        requested_queue_item_id=requested_queue_item_id,
         max_steps=11,
     )
